feat: image smoothing & blurring GUI

+++ src/gaussianFilter.m
@@ -1,3 +1,3 @@
 function img = gaussianFilter(image, filter_dim, sigma)
-    img = convolution(image, fspecial("gaussian", [filter_dim, filter_dim], sigma));
+    img = uint8(convolution(image, fspecial("gaussian", [filter_dim, filter_dim], sigma)));
 end
