+++ src/meanFilter.m
@@ -1,3 +1,3 @@
 function img = meanFilter(image, filter_dim)
-    img = convolution(image, (ones(filter_dim) / (filter_dim^2)));
+    img = uint8(convolution(image, (ones(filter_dim) / (filter_dim^2))));
 end
